// src/app/shared/models/Car.ts
export default interface Car {
  car_id: string;
  title: string;
  url: string;
  price: number;
  vin: string;
  year: string;
  mileage: string;
  description: string;
  city: string;
  region: string;
  created_time: string;
  photo: string;
  enginesize: string;
  engine_power: string;
  fuel_type: string;
  car_body: string;
  color: string;
  door_count: string;
  use_state: string;
  gearbox: string;
  steering_wheel: string;
  business: boolean;
  site_source: string;
}